Convert sensor nodes to catkin.
The sensor nodes does not depend on anything in the roburoc4 package
(though the magnetometer depends on Armadillo matrix lib). This
occured to me as I was porting the controllers, which will not compile
without the state estimator, which in turn depends on the sensors.
Thus you may see a commented section in the CMakeLists.txt pertaining
to the controllers, this will be uncommented later.

Signed-off-by: Karl Damkjær Hansen <kdh@es.aau.dk>