Add Mission Viewer
The Mission Supervisor was initially a hack for the Roedag demo,
it wasn't user friendly and had to be restarted after every misssion.

The new mission viewer uses Qt and aim at being more user friendly.
Still it is a hack for the Økologi Kongres demo.

This commit still needs to be field tested...

Signed-off-by: Karl Damkjær Hansen <kdh@es.aau.dk>